+++ Hierba-Santa/src/main/java/com/eon/hierbasanta/repository/TipoClienteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoClienteRepository extends JpaRepository<TipoCliente, Long> {
+public interface TipoClienteRepository extends JpaRepository<TipoCliente, Integer> {
 }
